+++ HISTORY.md
@@ -11,6 +11,7 @@
 * threads: added exception handling primitives for try/catch
 * extensions: added try-catch extension primitives
 * updated try-catch library
+* extensions: added object-naming extension primitive
 
 ### 2021-06-15
 * extensions: tweaked world-map primitives